+++ st2common/st2common/expressions/functions/json_escape.py
