+++ st2common/st2common/expressions/functions/jsonpath_query.py
