+++ st2common/st2common/expressions/functions/path.py
