+++ st2common/st2common/expressions/functions/regex.py
