+++ st2common/st2common/expressions/functions/time.py
