+++ st2common/st2common/expressions/functions/version.py
